+++ packages/graphql/lib/src/cache/cache.dart
@@ -120,7 +120,8 @@ class GraphQLCache extends NormalizingDataProxy {
     for (final patch in optimisticPatches) {
       if (patch.data.containsKey(rootId)) {
         final patchData = patch.data[rootId];
-        if (value is Map<String, Object> && patchData is Map<String, Object>) {
+        if (value is Map<String, dynamic> &&
+            patchData is Map<String, dynamic>) {
           value = deeplyMergeLeft([
             value,
             patchData,
@@ -204,9 +205,23 @@ class GraphQLCache extends NormalizingDataProxy {
   ///
   /// This allows for hierarchical optimism that is automatically cleaned up
   /// without having to tightly couple optimistic changes
+  ///
+  /// This is called on every network result as cleanup
   void removeOptimisticPatch(String removeId) {
+    final patchesToRemove = optimisticPatches
+        .where(
+          (patch) =>
+              patch.id == removeId || _parentPatchId(patch.id) == removeId,
+        )
+        .toList();
+
+    if (patchesToRemove.isEmpty) {
+      return;
+    }
+    // Only remove + mark broadcast requested if something was actually removed.
+    // This is to prevent unnecessary rebroadcasts
     optimisticPatches.removeWhere(
-      (patch) => patch.id == removeId || _parentPatchId(patch.id) == removeId,
+      (patch) => patchesToRemove.contains(patch),
     );
     broadcastRequested = true;
   }