+++ packages/graphql/lib/src/graphql_client.dart
@@ -26,11 +26,17 @@ class GraphQLClient implements GraphQLDataProxy {
     required this.cache,
     DefaultPolicies? defaultPolicies,
     bool alwaysRebroadcast = false,
+    DeepEqualsFn? deepEquals,
+    bool deduplicatePollers = false,
+    Duration? queryRequestTimeout = const Duration(seconds: 5),
   })  : defaultPolicies = defaultPolicies ?? DefaultPolicies(),
         queryManager = QueryManager(
           link: link,
           cache: cache,
           alwaysRebroadcast: alwaysRebroadcast,
+          deepEquals: deepEquals,
+          deduplicatePollers: deduplicatePollers,
+          requestTimeout: queryRequestTimeout,
         );
 
   /// The default [Policies] to set for each client action
@@ -45,16 +51,20 @@ class GraphQLClient implements GraphQLDataProxy {
   late final QueryManager queryManager;
 
   /// Create a copy of the client with the provided information.
-  GraphQLClient copyWith(
-      {Link? link,
-      GraphQLCache? cache,
-      DefaultPolicies? defaultPolicies,
-      bool? alwaysRebroadcast}) {
+  GraphQLClient copyWith({
+    Link? link,
+    GraphQLCache? cache,
+    DefaultPolicies? defaultPolicies,
+    bool? alwaysRebroadcast,
+    Duration? queryRequestTimeout,
+  }) {
     return GraphQLClient(
-        link: link ?? this.link,
-        cache: cache ?? this.cache,
-        defaultPolicies: defaultPolicies ?? this.defaultPolicies,
-        alwaysRebroadcast: alwaysRebroadcast ?? queryManager.alwaysRebroadcast);
+      link: link ?? this.link,
+      cache: cache ?? this.cache,
+      defaultPolicies: defaultPolicies ?? this.defaultPolicies,
+      alwaysRebroadcast: alwaysRebroadcast ?? queryManager.alwaysRebroadcast,
+      queryRequestTimeout: queryRequestTimeout ?? queryManager.requestTimeout,
+    );
   }
 
   /// This registers a query in the [QueryManager] and returns an [ObservableQuery]