+++ packages/graphql_flutter/lib/src/widgets/hooks/subscription.dart
@@ -69,8 +69,10 @@ class _SubscriptionHookState<TParsed> extends HookState<
     Stream<QueryResult<TParsed>>, _SubscriptionHook<TParsed>> {
   late Stream<QueryResult<TParsed>> stream;
 
-  ConnectivityResult? _currentConnectivityResult;
-  StreamSubscription<ConnectivityResult>? _networkSubscription;
+  List<ConnectivityResult> _currentConnectivityResult = [
+    ConnectivityResult.none
+  ];
+  StreamSubscription<List<ConnectivityResult>>? _networkSubscription;
 
   void _initSubscription() {
     final client = hook.client;
@@ -107,12 +109,12 @@ class _SubscriptionHookState<TParsed> extends HookState<
     super.dispose();
   }
 
-  Future<void> _onNetworkChange(ConnectivityResult result) async {
+  Future<void> _onNetworkChange(List<ConnectivityResult> results) async {
     //if from offline to online
-    if (_currentConnectivityResult == ConnectivityResult.none &&
-        (result == ConnectivityResult.mobile ||
-            result == ConnectivityResult.wifi)) {
-      _currentConnectivityResult = result;
+    if (_currentConnectivityResult.contains(ConnectivityResult.none) &&
+        (results.contains(ConnectivityResult.mobile) ||
+            results.contains(ConnectivityResult.wifi))) {
+      _currentConnectivityResult = List.from(results, growable: false);
 
       // android connectivitystate cannot be trusted
       // validate with nslookup
@@ -125,13 +127,13 @@ class _SubscriptionHookState<TParsed> extends HookState<
           }
           // on exception -> no real connection, set current state to none
         } on SocketException catch (_) {
-          _currentConnectivityResult = ConnectivityResult.none;
+          _currentConnectivityResult = [ConnectivityResult.none];
         }
       } else {
         _initSubscription();
       }
     } else {
-      _currentConnectivityResult = result;
+      _currentConnectivityResult = List.from(results, growable: false);
     }
   }